Corectare coordonate porti logice din float in integer

+++ Widgets/LogicGates/UI/NAND.py
@@ -37,7 +37,7 @@ GATE = {
         {
           "type": "circle",
           "x": 196,
-          "y": 196/2-5,
+          "y": int(196/2-5),
           "height": 20,
           "width": 20,
         },
